Improve region resolving
 - Share region resolving between SDK and CLI
 - Cache resolved region
 - Fail hard for CLI, fallback to default region for SDK
